Update Dockerfile
the huggingface models where to big when I included this so I'm going to move this change to a new branch "huggingface" so it can be included there and comment out most models in the manager file in the goldenverba as instructed by the original verba docs. This way the image file isn't too large to be uploaded on docker hub which is what I ran into with this.